requirement.txt
langchain
python-dotenv
langchain_community
langchain-text-splitters
langchain-openai
chromadb
langchain-chroma
sentence-transformers
langchain-huggingface
tqdm


langgraph
langchain-groq

pymupdf
fastmcp
fpdf
pywin32
langchain-mcp-adapters